Add test for inefficient encodings
The aggregate radix of repeating minimal labels should yield the least
efficient encoding. This test case proves that these can be decoded,
but it seems unlikely that most of these encodings are possible. In
order for the encoder to work, each legal input must have a bijective
encoding that decodes back to the original input value. If the encoder
indeed works, and this seems to be the case based on other tests, then,
with the exception of the "b.b..." encoding, this is clearly not the
case for these examples, since re-encoding the decoded output yields a
different result. The test case does, however, prove that these outputs
can be re-encoded again and then be successfully decoded. The length of
the initial decoded output gives a clue that the lower bound for inputs
that will fit into the 248 byte output may be 152 bytes.

Signed-off-by: Edwin Peer <espeer@gmail.com>